kisebb hibák javítása. és serilog debugok hozzádva a  klienshez -d kapcsoloval bekapcsolhato amugy warning szint megy

+++ C#/IktatogRPCClient/IktatogRPCClient/ViewModels/DocumentHandlerViewModel.cs
@@ -4,6 +4,7 @@ using Iktato;
 using IktatogRPCClient.Models;
 using IktatogRPCClient.Models.Managers;
 using Microsoft.Win32;
+using Serilog;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -23,6 +24,8 @@ namespace IktatogRPCClient.ViewModels
 			LoadData(IkonyvId);
 		}
 		private async void LoadData(int IkonyvId) {
+		
+			Log.Debug("{Class} Adatok betöltése a szerverről. Id: {IkonyvId}", GetType(),IkonyvId);
 			IkonyvDocuments = await serverHelper.GetDocumentInfoByIkonyvIdAsync(IkonyvId);
 		}
 		private int ikonyvid;
@@ -60,11 +63,17 @@ namespace IktatogRPCClient.ViewModels
 		}
 
 		public async void RemoveDocument() {
+			Log.Debug("{Class} Dokumentum törlése gomb megnyomva. Doc: {SelectedDocument}", GetType(), SelectedDocument);
 			if ( await serverHelper.RemoveDocumentAsync(SelectedDocument))
 			{
 				ModificationHappend = true;
 				IkonyvDocuments.Remove(SelectedDocument);
 				NotifyOfPropertyChange(() => IkonyvDocuments);
+				Log.Debug("{Class} Sikeres törlés", GetType());
+			}
+			else {
+
+				Log.Debug("{Class} Sikertelen törlés.", GetType());
 			}
 		}
 		public bool CanDownloadDocument {
@@ -78,43 +87,65 @@ namespace IktatogRPCClient.ViewModels
 		public async Task DownloadDocument() {
 		
 			LoaderIsVisible = true;
+			Log.Debug("{Class} Adatok letöltése a szerverről. Document: {SelectedDocument}", GetType(), SelectedDocument);
 			Document rawdata = await serverHelper.GetDocumentByIdAsync(SelectedDocument);
-			byte[] bytes = rawdata.Doc.ToByteArray();
-			string temppath = System.IO.Path.GetTempPath();
-			string fullpath = $"{temppath}{rawdata.Name}.{rawdata.Type}";
-			if (File.Exists(fullpath))
-			{
-				File.Delete(fullpath);
-				File.WriteAllBytes(fullpath, bytes);
+			if (rawdata.Id != -1) {
+				Log.Debug("{Class} Sikeres letöltés", GetType(), SelectedDocument);
+
+				byte[] bytes = rawdata.Doc.ToByteArray();
+				string temppath = Path.GetTempPath();
+				string fullpath = $"{temppath}{rawdata.Name}.{rawdata.Type}";
+				Log.Debug("{Class} Adatok mentése", GetType(), SelectedDocument);
+				if (File.Exists(fullpath))
+				{
+					File.Delete(fullpath);
+					File.WriteAllBytes(fullpath, bytes);
+				}
+				else File.WriteAllBytes(fullpath, bytes);
+				Log.Debug("{Class} Sikeres adatmentés.", GetType(), SelectedDocument);
+				try
+				{
+					Log.Debug("{Class} Letöltött documentum megnyitása.", GetType(), SelectedDocument);
+					Process.Start(fullpath);
+				}
+				catch (Exception e)
+				{
+					InformationBox.ShowError(e);
+				}
 			}
-			else File.WriteAllBytes(fullpath, bytes);
-		
-			try
-			{
-				Process.Start(fullpath);
-			}
-			catch (Exception e)
-			{
-				throw;
-			}
+			
 			LoaderIsVisible = false;
 		}
 		public async Task UploadDocument() {
+			Log.Debug("{Class} Feltöltés gomb megnyomva.", GetType());
+			Log.Debug("{Class} Adatok beszerzése", GetType());
 			string[] FileInfo = ChooseDataToUpload();
 			if (string.IsNullOrWhiteSpace(FileInfo[2])) return;
+			Log.Debug("{Class} Fájl feltöltés megkezdése.", GetType());
 			LoaderIsVisible = true;
 			Document document = new Document();
 			document.Name = FileInfo[0];
 			document.Type = FileInfo[1];
+			Log.Debug("{Class} ByteArray to BysteString", GetType());
 			document.Doc = ByteString.CopyFrom(GetBytesFromFile(FileInfo[2]));
 			document.IkonyvId = ikonyvid;
+			Log.Debug("{Class} Adatok feltöltése. Doc: {Name}", GetType(), document.Name);
 			DocumentInfo uploadedDocument = await serverHelper.UploadDocumentAsync(document);
-			IkonyvDocuments.Add(uploadedDocument);
-			NotifyOfPropertyChange(()=>IkonyvDocuments);
+			if (uploadedDocument.Id != -1)
+			{
+				Log.Debug("{Class} Sikeres feltöltése. ", GetType());
+				IkonyvDocuments.Add(uploadedDocument);
+				NotifyOfPropertyChange(() => IkonyvDocuments);
+
+				ModificationHappend = true;
+			}
+			else {
+				Log.Debug("{Class} Sikertelen feltöltése. ", GetType());
+			}
 			LoaderIsVisible = false;
-			ModificationHappend = true;
 		}
 		public void CancelButton() {
+			Log.Debug("{Class} Mégse gomb megnyomva. ", GetType());
 			eventAggregator.PublishOnUIThread(new DocumentHandlerClosed(ModificationHappend, IkonyvDocuments.Count > 0));
 			TryClose();
 		}
@@ -123,6 +154,7 @@ namespace IktatogRPCClient.ViewModels
 			string[] fileInfo = new string[3];
 			try
 			{
+				Log.Debug("{Class} Dialog megnyitása ", GetType());
 				OpenFileDialog dialog = new OpenFileDialog();
 				dialog.Title = "Válaszd ki a dokumentumot.";
 				dialog.FileName = "";
@@ -133,15 +165,18 @@ namespace IktatogRPCClient.ViewModels
 				{
 					string dokname = dialog.SafeFileName;
 					int lastDotIndex = dokname.LastIndexOf('.');
-
+					Log.Debug("{Class} Sikeres fájl kiválasztás. Name: {Dokname}", GetType(), dokname);
 					fileInfo[0] = dokname.Substring(0, lastDotIndex);
 					fileInfo[1] = dokname.Substring(lastDotIndex + 1);
 					fileInfo[2] = dialog.FileName;
 					return fileInfo;
 				}
 				else
-					
+				{
+					Log.Debug("{Class} Sikertelen fájl kiválasztás.", GetType());
 					return fileInfo;
+				}
+					
 			}
 			catch (Exception e) {
 				InformationBox.ShowError(e);
@@ -154,38 +189,41 @@ namespace IktatogRPCClient.ViewModels
 			BinaryReader br;
 			try
 			{
-
+				Log.Debug("{Class} Fájl megnyitása beolvasásra. ", GetType());
 				fs = new FileStream(FileName, FileMode.Open);
 				long lFileSize = fs.Length;
 				br = new BinaryReader(fs);
 				ba1 = br.ReadBytes((Int32)lFileSize);
-
+				Log.Debug("{Class} Fájl beolvasva. ", GetType());
 				br.Close();
 				fs.Close();
 			}
 			catch (UnauthorizedAccessException)
 			{
+				Log.Debug("{Class} Nincs jogosultság. A fájl átmásolása tempbe. ", GetType());
 				string temppath = Path.GetTempPath();
 				string fullpath = temppath + $"{FileName}";
 				if (File.Exists(fullpath))
 				{
+					Log.Debug("{Class} Ilyen nevu fájl már létezett. Törlése... ", GetType());
 					File.Delete(fullpath);
 					File.Copy(FileName, fullpath);
 				}
 				else File.Copy(FileName, fullpath);
+				Log.Debug("{Class} Fájl másolása sikeres. ", GetType());
 				fs = new FileStream(fullpath, FileMode.Open);
 				long lFileSize = fs.Length;
-
+				Log.Debug("{Class} Fájl beolvasása. ", GetType());
 				br = new BinaryReader(fs);
 				ba1 = br.ReadBytes((Int32)lFileSize);
-
+				Log.Debug("{Class} Fájl beolvasva. ", GetType());
 				br.Close();
 				fs.Close();
 
 			}
 			catch (Exception e)
 			{
-				System.Windows.MessageBox.Show(e.Message);
+				InformationBox.ShowError(e);
 			}
 			return (ba1);
 		}
